Move challenge v1 handlers into their own file
